+++ src/pages/SignIn/index.js
@@ -5,8 +5,10 @@ import { SignIn as SignInForm } from '~/views/Form';
 class SignIn extends Component {
   render() {
     return (
-      <Page center>
-        <SignInForm size="md" />
+      <Page>
+        <div className="d-flex justify-content-center">
+          <SignInForm size="md" />
+        </div>
       </Page>
     );
   }
